refactor: 인프라 빈 구성 Config를 분리 - @EnableMyAutoConfiguration : 별도의 메타 애노테이션에 @Import 애노테이션을 추가하도록 분리함. - 자동 생성&설정 필요한 클래스들을 autoconfig 디렉토리로이동
