+++ HelloBoot/src/main/java/tobyspring/config/EnableMyAutoConfiguration.java
@@ -1,0 +1,16 @@
+package tobyspring.config;
+
+import org.springframework.context.annotation.Import;
+import tobyspring.config.autoconfig.DispatcherServletConfig;
+import tobyspring.config.autoconfig.TomcatWebServerConfig;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.TYPE)
+@Import({DispatcherServletConfig.class, TomcatWebServerConfig.class})   // @Import 이용하여 스캔 영역에 없는 클래스를 스캔 및 구성 정보로 추가되도록 함.
+public @interface EnableMyAutoConfiguration {
+}
