+++ HelloBoot/src/main/java/tobyspring/config/MyComponent.java
@@ -1,4 +1,4 @@
-package tobyspring.helloboot;
+package tobyspring.config;
 
 import org.springframework.stereotype.Component;
