+++ HelloBoot/src/main/java/tobyspring/config/autoconfig/DispatcherServletConfig.java
@@ -1,0 +1,13 @@
+package tobyspring.config.autoconfig;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.web.servlet.DispatcherServlet;
+/**
+ * 자동 구성 정보에 해당할 - Servlet bean 설정 클래스
+ */
+public class DispatcherServletConfig {
+    @Bean
+    public DispatcherServlet dispatcherServlet() {
+        return new DispatcherServlet();
+    }
+}
