+++ HelloBoot/src/main/java/tobyspring/config/autoconfig/TomcatWebServerConfig.java
@@ -1,0 +1,15 @@
+package tobyspring.config.autoconfig;
+
+import org.springframework.boot.web.embedded.tomcat.TomcatServletWebServerFactory;
+import org.springframework.boot.web.servlet.server.ServletWebServerFactory;
+import org.springframework.context.annotation.Bean;
+
+/**
+ * 자동 구성 정보에 해당할 - 웹서버 bean 설정 클래스
+ */
+public class TomcatWebServerConfig {
+    @Bean
+    public ServletWebServerFactory servletWebServerFactory() {
+        return new TomcatServletWebServerFactory();
+    }
+}
